+++ ReleaseNotes/release_notes.md
@@ -4,6 +4,7 @@
   1. add `clenshaw-curtis` adaptive quadrature. It apprears to be better than Romberg' method.
   1. extend tests for Clenshaw-Curtis quadrature 
   1. Add modified PERT fuction and corresponding PDF
+  1. Add more methods to `Ostap::Math::Positive`, considering the positive polynomial like PDF 
 
 ## Backward incompatible: